requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
Pygments==2.14.0
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
rich==13.3.2
urllib3==1.26.14
